Move benchmark to separate resource set
